+++ Classes/FrontEnd/EventEditor.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
+use OliverKlee\Oelib\Email\GeneralEmailRole;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Utility\MailUtility;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
 /**
@@ -1420,7 +1422,12 @@ class Tx_Seminars_FrontEnd_EventEditor extends \Tx_Seminars_FrontEnd_Editor
             /** @var \Tx_Oelib_Mail $eMail */
             $eMail = GeneralUtility::makeInstance(\Tx_Oelib_Mail::class);
             $eMail->addRecipient($reviewer);
-            $eMail->setSender(self::getLoggedInUser());
+            $eMail->setSender(GeneralUtility::makeInstance(
+                GeneralEmailRole::class,
+                MailUtility::getSystemFromAddress(),
+                MailUtility::getSystemFromName()
+            ));
+            $eMail->setReplyTo(self::getLoggedInUser());
             $eMail->setSubject($this->translate('publish_event_subject'));
             $eMail->setMessage($this->createEMailContent($event));
 
@@ -1528,7 +1535,12 @@ class Tx_Seminars_FrontEnd_EventEditor extends \Tx_Seminars_FrontEnd_Editor
         /** @var \Tx_Oelib_Mail $eMail */
         $eMail = GeneralUtility::makeInstance(\Tx_Oelib_Mail::class);
         $eMail->addRecipient($reviewer);
-        $eMail->setSender(self::getLoggedInUser());
+        $eMail->setSender(GeneralUtility::makeInstance(
+            GeneralEmailRole::class,
+            MailUtility::getSystemFromAddress(),
+            MailUtility::getSystemFromName()
+        ));
+        $eMail->setReplyTo(self::getLoggedInUser());
         $eMail->setSubject($this->translate('save_event_subject'));
         $eMail->setMessage($this->createAdditionalEmailContent());